+++ src/blocks/column/block.js
@@ -4,7 +4,6 @@
 
 import classnames from "classnames"
 import UAGB_Block_Icons from "../../../dist/blocks/uagb-controls/block-icons"
-import styling from "./styling"
 import edit from "./edit"
 import save from "./save"
 import deprecated from "./deprecated"
@@ -15,30 +14,6 @@ import "./editor.scss"
 const { __ } = wp.i18n
 
 const { registerBlockType } = wp.blocks
-
-const {
-	InnerBlocks,
-	ColorPalette,
-	InspectorControls,
-	MediaUpload,
-	PanelColorSettings,
-} = wp.blockEditor
-
-const {
-	PanelBody,
-	RangeControl,
-	SelectControl,
-	Button,
-	ButtonGroup,
-	BaseControl,
-	TabPanel,
-	Dashicon
-} = wp.components
-
-const {
-	Component,
-	Fragment,
-} = wp.element
 
 registerBlockType( "uagb/column", {
 	title: uagb_blocks_info.blocks["uagb/column"]["title"],
