+++ src/blocks/columns/attributes.js
@@ -6,6 +6,10 @@ const attributes = {
 	block_id: {
 		type: "string",
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	columns: {
 		type: "number",
 		default: 2
