+++ src/blocks/columns/block.js
@@ -6,6 +6,7 @@ import classnames from "classnames"
 import UAGB_Block_Icons from "../../../dist/blocks/uagb-controls/block-icons"
 import attributes from "./attributes"
 import edit from "./edit"
+import deprecated from "./deprecated"
 import shapes from "./shapes"
 import "./style.scss"
 import "./editor.scss"
@@ -112,9 +113,9 @@ registerBlockType( "uagb/columns", {
 					`uagb-columns__gap-${columnGap}`,
 					`align${ align }`,
 					reverse_tablet,
-					reverse_mobile
+					reverse_mobile,
+					`uagb-block-${block_id}`
 				) }
-				id={ `uagb-columns-${block_id}` }
 			>
 				<div className="uagb-columns__overlay"></div>
 				{ top_divider_html }
@@ -138,104 +139,5 @@ registerBlockType( "uagb/columns", {
 			</CustomTag>
 		)
 	},
-	deprecated: [
-		{
-			attributes,			
-			save: function( props ) {
-
-				const { attributes, className } = props
-
-				const {
-					block_id,
-					tag,
-					backgroundType,
-					backgroundVideo,
-					contentWidth,
-					align,
-					columns,
-					stack,
-					vAlign,
-					columnGap,
-					topType,
-					bottomType,
-					bottomFlip,
-					topFlip,
-					reverseTablet,
-					reverseMobile,
-					topContentAboveShape,
-					bottomContentAboveShape
-				} = props.attributes
-
-				const CustomTag = `${tag}`
-
-				const top_divider_html = (
-					topType != "none" && (
-						<div
-							className={ classnames(
-								"uagb-columns__shape",
-								"uagb-columns__shape-top",
-								{ "uagb-columns__shape-flip": topFlip === true },
-								{ "uagb-columns__shape-above-content": topContentAboveShape === true }
-							) }>
-							{shapes[topType]}
-						</div>
-					)
-				)
-
-				const bottom_divider_html = (
-					bottomType != "none" && (
-						<div
-							className={ classnames(
-								"uagb-columns__shape",
-								"uagb-columns__shape-bottom",
-								{ "uagb-columns__shape-flip": bottomFlip === true },
-								{ "uagb-columns__shape-above-content": bottomContentAboveShape === true }
-							) }
-							data-negative="false">
-							{shapes[bottomType]}
-						</div>
-					)
-				)
-
-				const reverse_tablet = ( reverseTablet ) ? "uagb-columns__reverse-tablet" : ""
-
-				const reverse_mobile = ( reverseMobile ) ? "uagb-columns__reverse-mobile" : ""
-
-				return (
-					<CustomTag
-						className={ classnames(
-							className,
-							"uagb-columns__wrap",
-							`uagb-columns__background-${backgroundType}`,
-							`uagb-columns__stack-${stack}`,
-							`uagb-columns__valign-${vAlign}`,
-							`uagb-columns__gap-${columnGap}`,
-							`align${ align }`,
-							reverse_tablet,
-							reverse_mobile
-						) }
-						id={ `uagb-columns-${block_id}` }
-					>
-						<div className="uagb-columns__overlay"></div>
-						{ top_divider_html }
-						{ "video" == backgroundType &&
-							<div className="uagb-columns__video-wrap">
-								{  backgroundVideo &&
-									<video src={ backgroundVideo.url } autoPlay loop muted></video>
-								}
-
-							</div>
-						}
-						<div className={ classnames(
-							"uagb-columns__inner-wrap",
-							`uagb-columns__columns-${columns}`
-						) }>
-							<InnerBlocks.Content />
-						</div>
-						{ bottom_divider_html }
-					</CustomTag>
-				)
-			},
-		}
-	],
+	deprecated,
 } )
